feat: allow setting the ingress pathType value
Add a new value `ingress.pathType` to the datahub-frontend and datahub-gms to allow configuring the pathType parameter instead of the hard coded `ImplementationSpecific`. This is required to support ingress provided like GCE. Without this value, the implementationSpecific can behave in arbitrary ways, with GCE it behaves as an exact match and the UI can not load any assests, receiving a 404 instead.

Signed-off-by: Lucas Roesler <roesler.lucas@gmail.com>